pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "myapp"
version = "1.0.0"
description = "gangAI-labs api-gateway-ws-redis"
# readme = "README/README.md"  # Commented out - file doesn't exist
license = {text = "All Rights Reserved"}
authors = [
    {name = "gangAI-labs", email = "tomvsndev@gmail.com"},
]
requires-python = ">=3.10, <3.13"
keywords = ["fastapi", "pytorch", "production", "starter"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",  
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
]
dependencies = [
    "aiocircuitbreaker>=2.0.0",
    "colorlog>=6.9.0",
    "dotenv>=0.9.9",
    "fastapi>0.117",
    "httpx>=0.28.1",
    "python-decouple>=3.8",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "setproctitle>=1.3.7",
    "slowapi>=0.1.9",
    "tenacity>=9.0.0",
    "uvicorn>=0.37.0",
    "aioredis>=2.0.1",
    "redis>=6.4.0",
    "websockets>=15.0.1",
    "aiohttp>=3.12.15",
    "orjson>=3.11.3",
    "autobahn>=24.4.2",
    "uvloop>=0.21.0",
    "ujson>=5.11.0",
    "hypercorn>=0.17.3",
    "psutil>=7.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
 
[project.urls]
Homepage = "https://github.com/gangAI-labs/gangAI-labs"
Repository = "https://github.com/gangAI-labs/gangAI-labs.git"
Documentation = "https://github.com/yourusername/gangAI-labs#readme"
Issues = "https://github.com/yourusername/gangAI-labs/issues"

[project.scripts]
embedder-demo = "app:app"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?'
